Extract data + minor adjustment
Added a "py_scripts" folder where we can find the py script or the notebook that processed the data. The difference is that in the notebook there is an explorative approach.

Changed the structure of the "data" folder.

Some useless "id" referrences in the html are deleted. Others are substituted with "class" references.