Assignment1 page + first plot
Uploaded the assignment1.html file referring to the first assignment. The text  in it must be changed.

Inserted the first plot (BarChart of aboundance of each tree species) in assignment1.html using the js/assign1-plot1.js file.

Inserted a link in the index.html to assignment1.html

Changed some styles in css/styles.css and added the one required to plot the graph.

Future work:
- Write something in index.html for assignment1.
- Write description for the plot in assignment1.html.
- Add a link back to index.html.
- Possibly, do the point before using a sort of "navbar".
- Check if the plot looks good!