Extraction data for assign1 - plot 1 and 2
Modified the extraction of the data: now, it uses "geopandas" library, without the need of a hand-parsing of the original json files.

Extracted data for plot 1 (barchart of abundance of tree species) and plot 2 (stackchart per circoscrizione e top-5 (+ others) species - absolute value).

The name of csv file is the same of the js file for the plot.

The data file are well structured and also ordered to NOT perform any manipulation with d3js (we can discuss about it, but Python could be an optimal solution for its ease of use).

Future work:
- Complete the first plot.
- Do the second one.
- Pre-process data and do the remaining 3 plots.